Removed data collector from metamorphic relation
